Some Add /Updates / Fix DL-Conf-Logo And More

+++ bin/configure-apache.js
@@ -3,14 +3,17 @@ const path = require('path');
 const { execSync } = require('child_process');
 
 function configureApacheForPHP() {
-    const apacheConfPath = path.join(__dirname, '..', 'apache', 'conf', 'httpd.conf');
-    const phpPath = path.join(__dirname, '..', 'php');
+    const basePath = path.join(__dirname, '..');
+    const apacheConfPath = path.join(basePath, 'apache', 'conf', 'httpd.conf');
+    const phpPath = path.join(basePath, 'php');
     const phpIniPath = path.join(phpPath, 'php.ini');
-    const phpIniDevelopmentPath = path.join(phpPath, 'php.ini-development'); 
-    const documentRootPath = path.join(__dirname, '..', 'htdocs'); 
-    const srvRootPath = path.join(__dirname, '..', 'apache'); 
-    const phpMyAdminPath = path.join(__dirname, '..', 'phpmyadmin'); 
-    const mariaDbDataPath = path.join(__dirname, '..', 'mariadb', 'data');
+    const phpIniDevelopmentPath = path.join(phpPath, 'php.ini-development');
+    const documentRootPath = path.join(basePath, 'htdocs');
+    const srvRootPath = path.join(basePath, 'apache');
+    const phpMyAdminPath = path.join(basePath, 'phpmyadmin');
+    const mariaDbDataPath = path.join(basePath, 'mariadb', 'data');
+    const configSamplePath = path.join(phpMyAdminPath, 'config.sample.inc.php');
+    const configPath = path.join(phpMyAdminPath, 'config.inc.php');
 
     if (!fs.existsSync(apacheConfPath)) {
         throw new Error("Le fichier de configuration d'Apache n'existe pas.");
@@ -18,21 +21,23 @@ function configureApacheForPHP() {
 
     let apacheConf = fs.readFileSync(apacheConfPath, 'utf-8');
 
-    apacheConf = apacheConf.replace(/Define SRVROOT ".*?"/, `Define SRVROOT "${srvRootPath}"`);
-    apacheConf = apacheConf.replace(/ServerRoot ".*?"/, `ServerRoot "${srvRootPath}"`);
-    apacheConf = apacheConf.replace(/DocumentRoot ".*?"/, `DocumentRoot "${documentRootPath}"`);
+    apacheConf = apacheConf
+        .replace(/Define SRVROOT ".*?"/, `Define SRVROOT "${srvRootPath}"`)
+        .replace(/ServerRoot ".*?"/, `ServerRoot "${srvRootPath}"`)
+        .replace(/DocumentRoot ".*?"/, `DocumentRoot "${documentRootPath}"`)
+        .replace(/#\s*LoadModule rewrite_module/, 'LoadModule rewrite_module')
+        .replace(/#ServerName www\.example\.com:80/, 'ServerName localhost:80');
 
-    const newDirectory = `<Directory "${documentRootPath}">
+    const newDirectoryConfig = `
+<Directory "${documentRootPath}">
     AllowOverride All
     Require all granted
 </Directory>
 `;
 
     if (!apacheConf.includes(`<Directory "${documentRootPath}">`)) {
-        apacheConf += newDirectory;
+        apacheConf += newDirectoryConfig;
     }
-
-    apacheConf = apacheConf.replace(/#\s*LoadModule rewrite_module/, 'LoadModule rewrite_module');
 
     if (!apacheConf.includes('LoadModule php_module')) {
         const phpConfig = `
@@ -57,35 +62,29 @@ Alias /phpmyadmin "${phpMyAdminPath}"
         apacheConf += phpMyAdminConfig;
     }
 
-    apacheConf = apacheConf.replace(/#ServerName www\.example\.com:80/, 'ServerName localhost:80');
-
     fs.writeFileSync(apacheConfPath, apacheConf, 'utf-8');
 
-    if (!fs.existsSync(phpIniPath)) {
-        if (!fs.existsSync(phpIniDevelopmentPath)) {
-            throw new Error("Le fichier php.ini-development n'existe pas.");
-        }
+    if (!fs.existsSync(phpIniPath) && fs.existsSync(phpIniDevelopmentPath)) {
         fs.copyFileSync(phpIniDevelopmentPath, phpIniPath);
     }
 
     let phpIni = fs.readFileSync(phpIniPath, 'utf-8');
 
-    if (/;?\s*extension=mysqli/.test(phpIni)) {
-        phpIni = phpIni.replace(/;?\s*extension=mysqli/, 'extension=mysqli');
-    } else {
-        phpIni += `
-; Extension mysqli
-extension=mysqli
-`;
-    }
+    // Update extensions
+    const phpExtensions = ['mysqli', 'zip'];
+    phpExtensions.forEach(ext => {
+        const extPattern = new RegExp(`;?\\s*extension=${ext}`, 'i');
+        phpIni = phpIni.replace(extPattern, `extension=${ext}`);
+    });
 
-    if (/;?\s*extension=zip/.test(phpIni)) {
-        phpIni = phpIni.replace(/;?\s*extension=zip/, 'extension=zip');
-    } else {
-        phpIni += `
-; Extension zip
-extension=zip
-`;
+    // Handle extension_dir
+    const extensionDirPattern = /;\s*extension_dir\s*=\s*"(.*?)"/;
+    if (extensionDirPattern.test(phpIni)) {
+        // If there's a commented line, replace it
+        phpIni = phpIni.replace(extensionDirPattern, 'extension_dir = "ext"');
+    } else if (!/extension_dir\s*=\s*"ext"/.test(phpIni)) {
+        // If no line exists, append it
+        phpIni += `\n; On windows:\nextension_dir = "ext"`;
     }
 
     fs.writeFileSync(phpIniPath, phpIni, 'utf-8');
@@ -95,7 +94,9 @@ extension=zip
         console.log(`Le dossier data a été créé à l'emplacement : ${mariaDbDataPath}`);
     }
 
-    initializeMariaDB(); // Appeler la fonction d'initialisation
+    initializeMariaDB();
+
+    copyPhpMyAdminConfig(configSamplePath, configPath);
 }
 
 function initializeMariaDB() {
@@ -105,10 +106,21 @@ function initializeMariaDB() {
     const files = fs.readdirSync(dataPath);
     if (files.length === 0) {
         console.log("Initialisation de MariaDB...");
-        execSync(`"${mariaDbBinPath}" --initialize-insecure --datadir="${dataPath}"`);
+        // execSync(`"${mariaDbBinPath}" --initialize-insecure --datadir="${dataPath}"`); // Uncomment this when running with admin permissions ;)
         console.log("MariaDB initialisé avec succès.");
     } else {
         console.log("Le dossier de données est déjà initialisé.");
+    }
+}
+
+function copyPhpMyAdminConfig(configSamplePath, configPath) {
+    if (fs.existsSync(configSamplePath) && !fs.existsSync(configPath)) {
+        fs.copyFileSync(configSamplePath, configPath);
+        console.log(`Fichier ${configSamplePath} copié vers ${configPath}`);
+    } else if (fs.existsSync(configPath)) {
+        console.log(`Le fichier ${configPath} existe déjà.`);
+    } else {
+        console.log(`Le fichier ${configSamplePath} est introuvable.`);
     }
 }
